* [tests] more changes/improvements/experimentations with tests module * [colors] bug fixing, testing and improvements * [symbol] symbols are now compared by their values by default (via `SYMBOL_PROVIDER::compare`) * [random] renamed range-and-predicate-taking overloads to from `result()` to `result_if()` (index, iterator, element) * #include cleanups (with tests to make sure they can be included by themselves) * some fixes to compile more stuff on clang (still a few pieces missing with C++23)

+++ tests/templates_tests.cpp
@@ -136,7 +136,7 @@ TEST(templates_test, apply_to_nth_works)
 	EXPECT_TRUE(can_apply_to_nth<3>);
 	EXPECT_FALSE(can_apply_to_nth<4>);
 	EXPECT_FALSE(can_apply_to_nth<54>);
-	EXPECT_FALSE(can_apply_to_nth<-1>);
+	EXPECT_FALSE(can_apply_to_nth<(size_t)-1>);
 
 	bool called = false;
 	ghassanpl::apply_to_nth<0>([&](auto&& a) { called = true; }, UnCopyable{});
